requirements.txt
streamlit==1.32.0
newspaper3k==0.2.8
openai==1.12.0 